Arduino-python-piano/python_code/requirements.txt
pyserial~=3.4
paho-mqtt~=1.5.0
pynput